encoder/src/requirements.txt
ffmpeg-python
pymediainfo
filetype
